pyproject.toml
[tool.poetry]
name = "pyReel"
version = "0.1.0"
description = "pyReel is a Python service that converts media files to H.265 format. It uses `ffmpeg` to perform the conversion and only keeps the video if the output file is smaller than the input file."
authors = ["awhipp"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
ffmpeg-python = "^0.2.0"
pydantic = "^2.8.2"
uvicorn = "^0.30.5"
fastapi = "^0.112.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.8.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
